feat: merge query table view, states for unsaved/modified query tabs
