+++ package.json
@@ -24,6 +24,7 @@
   "dependencies": {
     "@electron-toolkit/preload": "^3.0.2",
     "@electron-toolkit/utils": "^4.0.0",
+    "@radix-ui/react-collapsible": "^1.1.12",
     "monaco-sql-languages": "^0.15.1",
     "mysql2": "^3.15.3",
     "pg": "^8.16.3"
