+++ src/main/saved-queries-storage.ts
@@ -2,7 +2,6 @@ import type { SavedQuery, SavedQueriesStorage } from '@common/types'
 import { app } from 'electron'
 import { promises as fs } from 'node:fs'
 import { dirname, join } from 'node:path'
-import { randomUUID } from 'node:crypto'
 
 const QUERIES_FILENAME = 'saved-queries.json'
 
@@ -62,6 +61,7 @@ const writeQueriesToDisk = async (queries: StoredQueriesStorage): Promise<void> 
 
 export const saveQuery = async (
   connectionId: string,
+  id: string,
   name: string,
   content: string
 ): Promise<SavedQuery> => {
@@ -69,7 +69,7 @@ export const saveQuery = async (
   const now = new Date()
 
   const query: SavedQuery = {
-    id: randomUUID(),
+    id,
     name,
     content,
     createdAt: now,
