+++ src/renderer/src/api/client.ts
@@ -130,8 +130,8 @@ export const dbdeskClient = {
     return getDbdesk().loadQueries(connectionId)
   },
 
-  async saveQuery(connectionId: string, name: string, content: string): Promise<SavedQuery> {
-    return getDbdesk().saveQuery(connectionId, name, content)
+  async saveQuery(connectionId: string, id: string, name: string, content: string): Promise<SavedQuery> {
+    return getDbdesk().saveQuery(connectionId, id, name, content)
   },
 
   async deleteQuery(connectionId: string, queryId: string): Promise<void> {
